+++ cli/src/core.rs
@@ -2,6 +2,7 @@
 pub(crate) mod ast;
 pub(crate) mod base_path;
 pub(crate) mod choices;
+pub(crate) mod client_sdk;
 pub(crate) mod command;
 pub(crate) mod database;
 pub(crate) mod docker;
@@ -33,7 +34,6 @@ pub(crate) mod sync;
 pub(crate) mod template;
 pub(crate) mod token;
 pub(crate) mod tsconfig;
-pub(crate) mod universal_sdk;
 pub(crate) mod version_check;
 pub(crate) mod vscode;
 pub(crate) mod watermark;