+++ cli/src/core/docker.rs
@@ -1689,7 +1689,7 @@ pub(crate) fn add_service_definition_to_docker_compose(
         ),
         format!("/{}/core/node_modules", manifest_data.app_name),
         format!("/{}/monitoring/node_modules", manifest_data.app_name),
-        format!("/{}/universal-sdk/node_modules", manifest_data.app_name),
+        format!("/{}/client-sdk/node_modules", manifest_data.app_name),
         format!("/{}/node_modules", manifest_data.app_name),
     ];