+++ cli/src/core/manifest/worker.rs
@@ -79,6 +79,13 @@ config_struct!(
 
         #[serde(skip_serializing, skip_deserializing)]
         pub(crate) is_iam_configured: bool,
+        #[serde(skip_serializing, skip_deserializing)]
+        pub(crate) is_billing_configured: bool,
+
+        #[serde(skip_serializing, skip_deserializing)]
+        pub(crate) is_request_cache_needed: bool,
+        #[serde(skip_serializing, skip_deserializing)]
+        pub(crate) is_type_needed: bool,
 
         #[serde(skip_serializing, skip_deserializing)]
         pub(crate) with_mappers: bool,
@@ -178,6 +185,27 @@ impl InitializableManifestConfig for WorkerManifestData {
                 }
                 return false;
             }),
+
+            is_billing_configured: self.projects.iter().any(|project_entry| {
+                if project_entry.name == "billing" {
+                    return true;
+                }
+                return false;
+            }),
+
+            is_request_cache_needed: worker_metadata
+                .infrastructure
+                .as_ref()
+                .map_or(false, |i| i.contains(&Infrastructure::Redis))
+                || project_entry
+                    .resources
+                    .as_ref()
+                    .map_or(false, |r| r.cache.is_some())
+                || self.projects.iter().any(|project_entry| {
+                    project_entry.name == "iam" || project_entry.name == "billing"
+                }),
+            // Workers always need type because WorkerConsumer uses type<>()
+            is_type_needed: true,
 
             // Default to false, will be set by CLI flag
             with_mappers: false,