+++ blueprint/implementations/billing/stripe/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @forklaunch/implementation-billing-stripe
 
+## 0.5.12
+
+### Patch Changes
+
+- minor typing updates
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@0.8.12
+  - @forklaunch/interfaces-billing@0.8.12
+
+## 0.5.11
+
+### Patch Changes
+
+- update package dependency versions
+- Updated dependencies
+  - @forklaunch/implementation-billing-base@0.8.11
+  - @forklaunch/interfaces-billing@0.8.11
+
 ## 0.5.10
 
 ### Patch Changes